requirements.txt
huggingface-hub==0.33.1
llama-index-core==0.12.44
llama-index-embeddings-huggingface==0.5.5
llama-index-instrumentation==0.2.0
llama-index-llms-groq==0.3.2
llama-index-llms-ollama==0.6.2
llama-index-llms-openai==0.4.7
llama-index-llms-openai-like==0.4.0
llama-index-readers-file==0.4.9
llama-index-vector-stores-chroma==0.4.2
llama-index-workflows==1.0.1
chromadb==1.0.13
pypdf==5.7.0
PyPDF2==3.0.1
pytesseract==0.3.13
pdf2image==1.17.0
transformers==4.53.0
sentence-transformers==4.1.0

